+++ acceptance/cases/migration/change_table_test.rb
@@ -86,7 +86,7 @@ module ActiveRecord
 
       def test_index_exists
         with_change_table do |t|
-          if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("3")
+          if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("3") && ActiveRecord::gem_version <= Gem::Version.create('7.0.4')
             @connection.expect :index_exists?, nil, [:delete_me, :bar, {}]
             t.index_exists?(:bar, {})
           else
@@ -98,7 +98,7 @@ module ActiveRecord
 
       def test_index_exists_with_options
         with_change_table do |t|
-          if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("3")
+          if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new("3") && ActiveRecord::gem_version <= Gem::Version.create('7.0.4')
             @connection.expect :index_exists?, nil, [:delete_me, :bar, {unique: true}]
             t.index_exists?(:bar, {unique: true})
           else