+++ lib/activerecord_spanner_adapter/information_schema.rb
@@ -79,6 +79,18 @@ module ActiveRecordSpannerAdapter
         column_name = row["COLUMN_NAME"]
         options = column_options[column_name]
 
+        default = row["COLUMN_DEFAULT"]
+        default_function = row["GENERATION_EXPRESSION"]
+
+        if /\w+\(.*\)/.match?(default)
+          default_function ||= default
+          default = nil
+        end
+
+        if default && type == "STRING"
+          default = unquote_string default
+        end
+
         Table::Column.new \
           table_name,
           column_name,
@@ -87,8 +99,8 @@ module ActiveRecordSpannerAdapter
           allow_commit_timestamp: options["allow_commit_timestamp"],
           ordinal_position: row["ORDINAL_POSITION"],
           nullable: row["IS_NULLABLE"] == "YES",
-          default: row["COLUMN_DEFAULT"],
-          default_function: row["GENERATION_EXPRESSION"],
+          default: default,
+          default_function: default_function,
           generated: row["GENERATION_EXPRESSION"].present?
       end
     end
@@ -278,7 +290,69 @@ module ActiveRecordSpannerAdapter
       [matched[1], limit]
     end
 
+    def unquote_string value
+      return unquote_raw_string value, 1 if value[0] == "r" || value[0] == "R"
+      unescape_string unquote_raw_string value
+    end
+
     private
+
+    def unquote_raw_string value, prefix_length = 0
+      triple_quote_range = prefix_length..(prefix_length + 2)
+      if value[triple_quote_range] == '"""' || value[triple_quote_range] == "'''"
+        value[(prefix_length + 3)...-3]
+      else
+        value[(prefix_length + 1)...-1]
+      end
+    end
+
+    def unescape_string value # rubocop:disable Metrics/AbcSize, Metrics/CyclomaticComplexity
+      str = ""
+      i = 0
+
+      while i < value.length
+        case value[i]
+        when "\\"
+          i += 1
+          case value[i]
+          when "a" then str += "\a"
+          when "b" then str += "\b"
+          when "f" then str += "\f"
+          when "n" then str += "\n"
+          when "r" then str += "\r"
+          when "t" then str += "\t"
+          when "v" then str += "\v"
+          when "\\" then str += "\\"
+          when "?" then str += "?"
+          when "'" then str += "'"
+          when '"' then str += '"'
+          when "`" then str += "`"
+          when "0".."7"
+            str += unescape_unicode value, i, 3, 8
+            i += 2
+          when "x", "X"
+            str += unescape_unicode value, i + 1, 2, 16
+            i += 2
+          when "u"
+            str += unescape_unicode value, i + 1, 4, 16
+            i += 4
+          when "U"
+            str += unescape_unicode value, i + 1, 8, 16
+            i += 8
+          end
+        else
+          str += value[i]
+        end
+
+        i += 1
+      end
+
+      str
+    end
+
+    def unescape_unicode value, start, length, base
+      [value[start...(start + length)].to_i(base)].pack "U"
+    end
 
     def column_options table_name, column_name
       sql = +"SELECT COLUMN_NAME, OPTION_NAME, OPTION_TYPE, OPTION_VALUE"