+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+### 1.4.4 (2023-09-06)
+
+#### Bug Fixes
+
+* Support for changes in Rails 7.0.7. ([#260](https://github.com/googleapis/ruby-spanner-activerecord/issues/260)) 
+
+### 1.4.3 (2023-06-09)
+
+#### Bug Fixes
+
+* unquote string default value ([#253](https://github.com/googleapis/ruby-spanner-activerecord/issues/253)) 
+
+### 1.4.2 (2023-06-01)
+
+#### Bug Fixes
+
+* allow functions to be default values ([#252](https://github.com/googleapis/ruby-spanner-activerecord/issues/252)) 
+* use original types for composite primary keys ([#246](https://github.com/googleapis/ruby-spanner-activerecord/issues/246)) 
+
 ### 1.4.1 (2023-03-01)
 
 #### Bug Fixes